feat!: expose install version
