+++ src/index.js
@@ -1,3 +1,4 @@
 /* @flow */
 
 export { npmPoll as poll } from './poll';
+export { installVersion } from './npm';
